# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Newsight API
  description: REST API documentation for the Newsight backend.
  contact:
    name: Newsight
  version: v1
servers:
- url: http://localhost:8080
  description: Local backend
- url: https://newsightkr.com
  description: Production
...
paths:
  /api/admins/inquiries/{inquiry_seq}/answer/:
    put:
      tags:
      - Admin Inquiries
      summary: Save or update admin answer
      operationId: saveOrUpdateAdminAnswer
      parameters:
      - name: inquiry_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInquiryAnswerRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInquiryAnswerResponseDto"
      security:
      - bearerAuth: []
  /api/admins/inquiries/{inquiry_seq}/answer:
    put:
      tags:
      - Admin Inquiries
      summary: Save or update admin answer
      operationId: saveOrUpdateAdminAnswer_1
      parameters:
      - name: inquiry_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInquiryAnswerRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInquiryAnswerResponseDto"
      security:
      - bearerAuth: []
...
  /api/analytics/media-compare/keywords/{keyword_seq}/sentiment/content/:
    get:
      tags:
      - Media Compare Analytics
      summary: Get content sentiment by media
      operationId: getMediaCompareContentSentiment
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MediaContentSentimentCompareResponseDto"
  /api/analytics/media-compare/keywords/{keyword_seq}/sentiment/content:
    get:
      tags:
      - Media Compare Analytics
      summary: Get content sentiment by media
      operationId: getMediaCompareContentSentiment_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MediaContentSentimentCompareResponseDto"
  /api/analytics/media-compare/keywords/{keyword_seq}/media-article-counts:
    get:
      tags:
      - Media Compare Analytics
      summary: Get article counts by media
      operationId: getMediaArticleCounts
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MediaArticleCountsResponseDto"
  /api/analytics/media-compare/keywords/{keyword_seq}/media-article-counts/:
    get:
      tags:
      - Media Compare Analytics
      summary: Get article counts by media
      operationId: getMediaArticleCounts_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MediaArticleCountsResponseDto"
  /api/analytics/media-compare/keywords/{keyword_seq}/framing/title-top-words/:
    get:
      tags:
      - Media Compare Analytics
      summary: Get title top words by media
      operationId: getMediaCompareTitleTopWords
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      - name: top_n
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MediaTitleTopWordsResponseDto"
  /api/analytics/media-compare/keywords/{keyword_seq}/framing/title-top-words:
    get:
      tags:
      - Media Compare Analytics
      summary: Get title top words by media
      operationId: getMediaCompareTitleTopWords_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      - name: top_n
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MediaTitleTopWordsResponseDto"
  /api/analytics/media-compare/keywords/top/:
    get:
      tags:
      - Media Compare Analytics
      summary: List top keywords for media comparison
      operationId: getMediaCompareTopKeywords
      parameters:
      - name: period
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MediaCompareTopKeywordsResponseDto"
  /api/analytics/media-compare/keywords/top:
    get:
      tags:
      - Media Compare Analytics
      summary: List top keywords for media comparison
      operationId: getMediaCompareTopKeywords_1
      parameters:
      - name: period
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MediaCompareTopKeywordsResponseDto"
...
  /api/admins/inquiries/{inquiry_seq}:
    get:
      tags:
      - Admin Inquiries
      summary: Get inquiry detail for admin
      operationId: getInquiryDetailForAdmin
      parameters:
      - name: inquiry_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInquiryDetailResponseDto"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Admin Inquiries
      summary: Delete inquiry
      operationId: deleteInquiryForAdmin
      parameters:
      - name: inquiry_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInquiryDeleteResponseDto"
      security:
      - bearerAuth: []
  /api/admins/inquiries/{inquiry_seq}/:
    get:
      tags:
      - Admin Inquiries
      summary: Get inquiry detail for admin
      operationId: getInquiryDetailForAdmin_1
      parameters:
      - name: inquiry_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInquiryDetailResponseDto"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Admin Inquiries
      summary: Delete inquiry
      operationId: deleteInquiryForAdmin_1
      parameters:
      - name: inquiry_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInquiryDeleteResponseDto"
      security:
      - bearerAuth: []
  /api/admins/inquiries/:
    get:
      tags:
      - Admin Inquiries
      summary: List inquiries for admin
      operationId: listInquiriesForAdmin
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInquiryListResponseDto"
      security:
      - bearerAuth: []
  /api/admins/inquiries:
    get:
      tags:
      - Admin Inquiries
      summary: List inquiries for admin
      operationId: listInquiriesForAdmin_1
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminInquiryListResponseDto"
      security:
      - bearerAuth: []
...
components:
  schemas:
    AdminInquiryAnswerRequestDto:
      type: object
      properties:
        admin_message:
          type: string
          minLength: 1
        status:
          type: string
          minLength: 1
      required:
      - admin_message
      - status
    AdminInquiryAnswerResponseDto:
      type: object
      properties:
        inquiry_seq:
          type: integer
          format: int64
        status:
          type: string
        processed_at:
          type: string
          format: date-time
        answer_updated_at:
          type: string
          format: date-time
...
    MediaContentSentimentCompareResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaSentimentItemDto"
    MediaSentimentItemDto:
      type: object
      properties:
        media_name:
          type: string
        positive:
          type: number
          format: double
        neutral:
          type: number
          format: double
        negative:
          type: number
          format: double
    MediaArticleCountItemDto:
      type: object
      properties:
        media_name:
          type: string
        article_count:
          type: integer
          format: int32
    MediaArticleCountsResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaArticleCountItemDto"
    MediaTitleTopWordsResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaTopWordsItemDto"
    MediaTopWordsItemDto:
      type: object
      properties:
        media_name:
          type: string
        words:
          type: array
          items:
            type: string
    KeywordPillItemDto:
      type: object
      properties:
        keyword_seq:
          type: integer
          format: int64
        keyword:
          type: string
    MediaCompareTopKeywordsResponseDto:
      type: object
      properties:
        period_start:
          type: string
        period_end:
          type: string
        selected_keyword:
          type: string
        selected_article_count:
          type: integer
          format: int32
        selected_media_count:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/KeywordPillItemDto"
        selected_keyword_seq:
          type: integer
          format: int64
...
    AdminInquiryDetailResponseDto:
      type: object
      properties:
        inquiry:
          $ref: "#/components/schemas/AdminInquiryDto"
    AdminInquiryDto:
      type: object
      properties:
        inquiry_seq:
          type: integer
          format: int64
        type_code:
          type: string
        title:
          type: string
        message:
          type: string
        inquirer_id:
          type: string
        submitted_at:
          type: string
          format: date-time
        status:
          type: string
        admin_message:
          type: string
    AdminInquiryListItemDto:
      type: object
      properties:
        inquiry_seq:
          type: integer
          format: int64
        type_code:
          type: string
        title:
          type: string
        inquirer_id:
          type: string
        submitted_at:
          type: string
          format: date-time
        status:
          type: string
    AdminInquiryListResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminInquiryListItemDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total_count:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
...
    AdminInquiryDeleteResponseDto:
      type: object
      properties:
        inquiry_seq:
          type: integer
          format: int64
  securitySchemes:
    bearerAuth:
      type: http
      description: Use the access token returned by the login API.
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT